Adding support for gminer decimal hash rates

+++ foreman-gminer/src/main/java/mn/foreman/gminer/json/Stat.java
@@ -3,6 +3,7 @@ package mn.foreman.gminer.json;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -120,7 +121,7 @@ public class Stat {
 
         /** The speed. */
         @JsonProperty("speed")
-        public int speed;
+        public BigDecimal speed;
 
         /** The temperature. */
         @JsonProperty("temperature")
